refactor: create individual viewer dir
